+++ SecureHttpClient.TestRunner.Android/MainActivity.cs
@@ -1,5 +1,7 @@
 ﻿using Android.App;
 using Android.OS;
+using Serilog;
+using Serilog.Core;
 using Xunit.Runners.UI;
 
 namespace SecureHttpClient.TestRunner.Android
@@ -9,9 +11,21 @@ namespace SecureHttpClient.TestRunner.Android
     {
         protected override void OnCreate(Bundle bundle)
         {
+            Log.Logger = new LoggerConfiguration()
+                .MinimumLevel.Information()
+                .WriteTo.AndroidLog()
+                .Enrich.WithProperty(Constants.SourceContextPropertyName, "TestRunner")
+                .CreateLogger();
+
             AddTestAssembly(typeof(Test.SslTest).Assembly);
             AutoStart = true;
             base.OnCreate(bundle);
+        }
+
+        protected override void OnStop()
+        {
+            Log.CloseAndFlush();
+            base.OnStop();
         }
     }
 }