+++ SecureHttpClient.TestRunner.iOS/AppDelegate.cs
@@ -1,5 +1,7 @@
 ﻿using Foundation;
 using UIKit;
+using Serilog;
+using Serilog.Core;
 using Xunit.Runner;
 using Xunit.Sdk;
 
@@ -10,10 +12,22 @@ namespace SecureHttpClient.TestRunner.iOS
     {
         public override bool FinishedLaunching(UIApplication app, NSDictionary options)
         {
+            Log.Logger = new LoggerConfiguration()
+                .MinimumLevel.Information()
+                .WriteTo.NSLog()
+                .Enrich.WithProperty(Constants.SourceContextPropertyName, "TestRunner")
+                .CreateLogger();
+
             AddExecutionAssembly(typeof(ExtensibilityPointFactory).Assembly);
             AddTestAssembly(typeof(Test.SslTest).Assembly);
             AutoStart = true;
             return base.FinishedLaunching(app, options);
+        }
+
+        public override void WillTerminate(UIApplication app)
+        {
+            Log.CloseAndFlush();
+            base.WillTerminate(app);
         }
     }
 }